pyproject.toml
[project]
name = "content-summarization"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "ipykernel>=7.1.0",
    "langchain>=1.2.3",
    "langchain-anthropic>=1.3.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-google-genai>=4.1.3",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.7",
    "langchain-yt-dlp>=0.0.8",
    "python-dotenv>=1.2.1",
    "pytube>=15.0.0",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.52.2",
    "transformers>=4.57.3",
    "unstructured>=0.18.27",
    "validators>=0.35.0",
    "youtube-transcript-api>=1.2.3",
]
